api/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
supabase
anthropic
voyageai
httpx
python-dotenv
